feat: finish create command

+++ src/cli/create.ts
@@ -1,9 +1,11 @@
 
 import { Command, flags } from '@oclif/command'
 import { ServiceAPI, ServiceAPIMode } from '../api'
+import { cli } from 'cli-ux'
 import * as path from 'path'
+import ListCommand from './list'
 
-export default class StartCommand extends Command {
+export default class CreateCommand extends Command {
   static description = 'register your application to the init system and run it'
 
   static flags = {
@@ -11,8 +13,8 @@ export default class StartCommand extends Command {
     interpreter: flags.string({
       description: 'interpreter to use when launching your script (either binary name or absolute path to it)'
     }),
-    system: flags.boolean({
-      description: 'connect to the root init system (means that your process will be started as root)'
+    as: flags.string({
+      description: 'Choose permission to assign to the service (either user (default), nobody or root)'
     })
   }
 
@@ -25,13 +27,18 @@ export default class StartCommand extends Command {
   ]
 
   async run () {
-    const { args, flags } = this.parse(StartCommand)
-    const api = await ServiceAPI.init(flags.system ? ServiceAPIMode.SYSTEM : ServiceAPIMode.USER)
+    const { args, flags } = this.parse(CreateCommand)
     const scriptPath = path.resolve(process.cwd(), args.filename)
-    await api.create({
+    const api = await ServiceAPI.init(ServiceAPIMode[(flags.as || 'user').toUpperCase()])
+    if (process.getuid() !== 0) {
+      throw new Error(`You must use sudo with servicectl for it to work properly.`)
+    }
+
+    const service = await api.create({
       script: scriptPath,
       interpreter: flags.interpreter
     })
+    cli.table([ service ], ListCommand.headers)
     await api.destroy()
   }
 }
